+++ lib/opal/nodes/def.rb
@@ -1,26 +1,12 @@
-require 'opal/nodes/scope'
+require 'opal/nodes/node_with_args'
 
 module Opal
   module Nodes
     # FIXME: needs rewrite
-    class DefNode < ScopeNode
+    class DefNode < NodeWithArgs
       handle :def
 
       children :recvr, :mid, :args, :stmts
-
-      def opt_args
-        @opt_args ||= args[1..-1].select { |arg| arg.first == :optarg }
-      end
-
-      def rest_arg
-        @rest_arg ||= args[1..-1].find { |arg| arg.first == :restarg }
-      end
-
-      def keyword_args
-        @keyword_args ||= args[1..-1].select do |arg|
-          [:kwarg, :kwoptarg, :kwrestarg].include? arg.first
-        end
-      end
 
       def block_arg
         @block_arg ||= args[1..-1].find { |arg| arg.first == :blockarg }
@@ -135,113 +121,6 @@ module Opal
         wrap '(', ", nil) && '#{mid}'" if expr?
       end
 
-      def compile_block_arg
-        if scope.uses_block?
-          scope_name  = scope.identity
-          yielder     = scope.block_name
-
-          add_temp "$iter = #{scope_name}.$$p"
-          add_temp "#{yielder} = $iter || nil"
-
-          line "#{scope_name}.$$p = null;"
-        end
-      end
-
-      def compile_rest_arg
-        if rest_arg and rest_arg[1]
-          rest_var = variable(rest_arg[1].to_sym)
-          add_temp '$rest_idx'
-          add_temp "$rest_len = arguments.length - #{argc}"
-
-          line "var #{rest_var} = new Array($rest_len > 0 ? $rest_len : 0);"
-          line "if ($rest_len > 0) {"
-          indent do
-            line "for ($rest_idx = 0; $rest_idx < $rest_len; $rest_idx++) {"
-            line "  #{rest_var}[$rest_idx] = arguments[$rest_idx + #{argc}];"
-            line "}"
-          end
-          line '}'
-        end
-      end
-
-      def compile_opt_args
-        opt_args.each do |arg|
-          next if arg[2][2] == :undefined
-          line "if (#{variable(arg[1])} == null) {"
-          line "  #{variable(arg[1])} = ", expr(arg[2])
-          line "}"
-        end
-      end
-
-      def compile_keyword_args
-        return if keyword_args.empty?
-        helper :hash2
-
-        if rest_arg
-          with_temp do |tmp|
-            rest_arg_name = variable(rest_arg[1].to_sym)
-            line "#{tmp} = #{rest_arg_name}[#{rest_arg_name}.length - 1];"
-            line "if (#{tmp} == null || !#{tmp}.$$is_hash) {"
-            line "  $kwargs = $hash2([], {});"
-            line "} else {"
-            line "  $kwargs = #{rest_arg_name}.pop();"
-            line "}"
-          end
-        elsif last_opt_arg = opt_args.last
-          opt_arg_name = variable(last_opt_arg[1])
-          line "if (#{opt_arg_name} == null) {"
-          line "  $kwargs = $hash2([], {});"
-          line "}"
-          line "else if (#{opt_arg_name}.$$is_hash) {"
-          line "  $kwargs = #{opt_arg_name};"
-          line "  #{opt_arg_name} = ", expr(last_opt_arg[2]), ";"
-          line "}"
-          line "else if ($kwargs == null) {"
-          line "  $kwargs = $hash2([], {});"
-          line "}"
-        else
-          line "if ($kwargs == null) {"
-          line "  $kwargs = $hash2([], {});"
-          line "}"
-        end
-
-        line "if (!$kwargs.$$is_hash) {"
-        line "  throw Opal.ArgumentError.$new('expecting keyword args');"
-        line "}"
-
-        keyword_args.each do |kwarg|
-          case kwarg.first
-          when :kwoptarg
-            arg_name = kwarg[1]
-            var_name = variable(arg_name.to_s)
-            add_local var_name
-            line "if ((#{var_name} = $kwargs.$$smap['#{arg_name}']) == null) {"
-            line "  #{var_name} = ", expr(kwarg[2])
-            line "}"
-          when :kwarg
-            arg_name = kwarg[1]
-            var_name = variable(arg_name.to_s)
-            add_local var_name
-            line "if ((#{var_name} = $kwargs.$$smap['#{arg_name}']) == null) {"
-            line "  throw new Error('expecting keyword arg: #{arg_name}')"
-            line "}"
-          when :kwrestarg
-            arg_name = kwarg[1]
-            var_name = variable(arg_name.to_s)
-            add_local var_name
-
-            kwarg_names = keyword_args.select do |kw|
-              [:kwoptarg, :kwarg].include? kw.first
-            end.map { |kw| "#{kw[1].to_s.inspect}: true" }
-
-            used_args = "{#{kwarg_names.join ','}}"
-            line "#{var_name} = Opal.kwrestargs($kwargs, #{used_args});"
-          else
-            raise "unknown kwarg type #{kwarg.first}"
-          end
-        end
-      end
-
       # Returns code used in debug mode to check arity of method call
       def arity_check(args, opt, splat, kwargs, block_name, mid)
         meth = mid.to_s.inspect
@@ -280,13 +159,17 @@ module Opal
       def compile
         done_kwargs = false
         have_rest   = false
+        first_arg = true
 
         children.each_with_index do |child, idx|
           case child.first
           when :kwarg, :kwoptarg, :kwrestarg
-            unless done_kwargs
+            if have_rest
+              scope.args_after_rest_args << child
+            elsif !done_kwargs
               done_kwargs = true
-              push ', ' unless idx == 0 || have_rest
+              push ', ' unless first_arg
+              first_arg = false
               scope.add_arg '$kwargs'
               push '$kwargs'
             end
@@ -297,12 +180,40 @@ module Opal
           when :restarg
             have_rest = true
 
+            if idx == children.length - 1
+              push ', ' unless first_arg
+              push '$restarg'
+            end
+          when :mlhs
+
+            if have_rest
+              scope.args_after_rest_args << child
+            else
+              tmp = scope.next_temp
+              scope.add_arg tmp
+              push ', ' unless first_arg
+              first_arg = false
+              push tmp
+              scope.mlhs_mapping[child] = tmp
+
+              child.children.each do |child|
+                scope.add_temp variable(child.last)
+              end
+            end
+
           else
-            child = child[1].to_sym
-            push ', ' unless idx == 0
-            child = variable(child)
-            scope.add_arg child.to_sym
-            push child.to_s
+            if have_rest
+              # to handle cases like
+              # def m(a, *b, c, d: 1) -> function(a) { // extracting args manually }
+              scope.args_after_rest_args << child
+            else
+              child = child[1].to_sym
+              push ', ' unless first_arg
+              first_arg = false
+              child = variable(child.to_sym)
+              scope.add_arg child.to_sym
+              push child.to_s
+            end
           end
         end
       end