+++ spec/lib/parser/iter_spec.rb
@@ -21,39 +21,39 @@ describe "Iters" do
   end
 
   describe "with normal args" do
-    it "adds a single s(:lasgn) for 1 norm arg" do
-      parsed("proc do |a|; end")[4][1].should == [:lasgn, :a]
+    it "adds a single s(:masgn) for 1 norm arg" do
+      parsed("proc do |a|; end")[4][1].should == [:masgn, [:args, [:arg, :a]]]
     end
 
     it "lists multiple norm args inside a s(:masgn)" do
-      parsed("proc do |a, b|; end")[4][1].should == [:masgn, [:array, [:lasgn, :a], [:lasgn, :b]]]
-      parsed("proc do |a, b, c|; end")[4][1].should == [:masgn, [:array, [:lasgn, :a], [:lasgn, :b], [:lasgn, :c]]]
+      parsed("proc do |a, b|; end")[4][1].should == [:masgn, [:args, [:arg, :a], [:arg, :b]]]
+      parsed("proc do |a, b, c|; end")[4][1].should == [:masgn, [:args, [:arg, :a], [:arg, :b], [:arg, :c]]]
     end
   end
 
   describe "with splat arg" do
-    it "adds a s(:masgn) for the s(:splat) even if its the only arg" do
-      parsed("proc do |*a|; end")[4][1].should == [:masgn, [:array, [:splat, [:lasgn, :a]]]]
-      parsed("proc do |a, *b|; end")[4][1].should == [:masgn, [:array, [:lasgn, :a], [:splat, [:lasgn, :b]]]]
+    it "adds a s(:masgn) for the s(:restarg) even if its the only arg" do
+      parsed("proc do |*a|; end")[4][1].should == [:masgn, [:args, [:restarg, :a]]]
+      parsed("proc do |a, *b|; end")[4][1].should == [:masgn, [:args, [:arg, :a], [:restarg, :b]]]
     end
   end
 
   describe "with opt args" do
-    it "adds a s(:block) arg to end of s(:masgn) for each lasgn" do
-      parsed("proc do |a = 1|; end")[4][1].should == [:masgn, [:array, [:lasgn, :a], [:block, [:lasgn, :a, [:int, 1]]]]]
-      parsed("proc do |a = 1, b = 2|; end")[4][1].should == [:masgn, [:array, [:lasgn, :a], [:lasgn, :b], [:block, [:lasgn, :a, [:int, 1]], [:lasgn, :b, [:int, 2]]]]]
+    it "adds a s(:optarg) arg each optional argument" do
+      parsed("proc do |a = 1|; end")[4][1].should == [:masgn, [:args, [:optarg, :a, [:int, 1]]]]
+      parsed("proc do |a = 1, b = 2|; end")[4][1].should == [:masgn, [:args, [:optarg, :a, [:int, 1]], [:optarg, :b, [:int, 2]]]]
     end
 
-    it "should add lasgn block after all other args" do
-      parsed("proc do |a, b = 1|; end")[4][1].should == [:masgn, [:array, [:lasgn, :a], [:lasgn, :b], [:block, [:lasgn, :b, [:int, 1]]]]]
-      parsed("proc do |b = 1, *c|; end")[4][1].should == [:masgn, [:array, [:lasgn, :b], [:splat, [:lasgn, :c]], [:block, [:lasgn, :b, [:int, 1]]]]]
-      parsed("proc do |b = 1, &c|; end")[4][1].should == [:masgn, [:array, [:lasgn, :b], [:block_pass, [:lasgn, :c]], [:block, [:lasgn, :b, [:int, 1]]]]]
+    it "should add args in the direct order" do
+      parsed("proc do |a, b = 1|; end")[4][1].should == [:masgn, [:args, [:arg, :a], [:optarg, :b, [:int, 1]]]]
+      parsed("proc do |b = 1, *c|; end")[4][1].should == [:masgn, [:args, [:optarg, :b, [:int, 1]], [:restarg, :c]]]
+      parsed("proc do |b = 1, &c|; end")[4][1].should == [:masgn, [:args, [:optarg, :b, [:int, 1]], [:block_pass, [:lasgn, :c]]]]
     end
   end
 
   describe "with block arg" do
     it "should add block arg with s(:block_pass) wrapping s(:lasgn) prefix" do
-      parsed("proc do |&a|; end")[4][1].should == [:masgn, [:array, [:block_pass, [:lasgn, :a]]]]
+      parsed("proc do |&a|; end")[4][1].should == [:masgn, [:args, [:block_pass, [:lasgn, :a]]]]
     end
   end
 end